add source map controller
